Move trim_zeroes to misc_ui
