+++ src/GUI/OptionsGroup/Field.hpp
@@ -342,15 +342,7 @@ private:
     void _set_value(Slic3r::Pointf3 value);
     void _set_value(std::string value);
 
-    /// Remove extra zeroes generated from std::to_string on doubles
-    std::string trim_zeroes(std::string in) {
-        std::string result {""};
-        std::regex strip_zeroes("(0*)$");
-        std::regex_replace (std::back_inserter(result), in.begin(), in.end(), strip_zeroes, "");
-        if (result.back() == '.') result.append("0");
-        return result;
     }
-    wxString trim_zeroes(wxString in) { return wxString(trim_zeroes(in.ToStdString())); }
