+++ src/GUI/misc_ui.cpp
@@ -5,6 +5,7 @@
 
 #include <exception>
 #include <stdexcept>
+#include <regex>
 
 
 namespace Slic3r { namespace GUI {
@@ -146,6 +147,17 @@ std::vector<wxString> open_model(wxWindow* parent, wxWindow* top) {
     dialog->Destroy(); 
     return tmp;
 }
+
+/// Remove extra zeroes generated from std::to_string on doubles
+std::string trim_zeroes(std::string in) {
+    std::string result {""};
+    std::regex strip_zeroes("(0*)$");
+    std::regex_replace (std::back_inserter(result), in.begin(), in.end(), strip_zeroes, "");
+    if (result.back() == '.') result.append("0");
+    return result;
+}
+
+wxString trim_zeroes(wxString in) { return wxString(trim_zeroes(in.ToStdString())); }
 
 }} // namespace Slic3r::GUI
